convert the extension to my own extension most likely - for file upload as unlikely any use to others.
still need to add column to the file_column - or perhaps just the column table to indicate whether the files are public or private.

Add html purifier on save of html fields - in beforeSave most likely for ActiveRecord

Convert to all to psr i.e. replace tabs with spaces and clean up code with block comments etc